requirements.txt
django-annoying
django_enumfield>=2.0.0
djangorestframework
jsonschema
drf-extensions
toolz
psycopg2-binary

coverage
Faker
django-annoying
responses
djangorestframework-simplejwt
hashids==1.2.0
freezegun
twine
beautifultable
pyfcm==1.4.7
urllib3==1.26.0
